+++ src/main/java/Commands/AddUnmark.java
@@ -1,0 +1,23 @@
+package Commands;
+
+import WindeBot.Reminder;
+import WindeBot.Ui;
+
+import Exceptions.EmptyDescriptionException;
+import Exceptions.TooManyParametersException;
+
+public class AddUnmark extends Command {
+    public boolean execute(String input, Reminder reminder, Ui ui) throws EmptyDescriptionException, TooManyParametersException {
+        String[] command = input.split(" ");
+        if (command.length == 2) {
+            ui.print("OK, I've marked this task as not done yet:");
+            reminder.unmark(Integer.parseInt(command[1]) - 1);
+            ui.print("    " + reminder.getTask(Integer.parseInt(command[1]) - 1).toString());
+        } else if (command.length < 2) {
+            throw new EmptyDescriptionException("I NEED TO KNOW WHAT I'M MARKING!");
+        } else {
+            throw new TooManyParametersException("ONE AT A TIME!");
+        }
+        return true;
+    }
+}
