+++ src/main/java/Commands/ByeCommand.java
@@ -1,0 +1,11 @@
+package Commands;
+
+import WindeBot.Reminder;
+import WindeBot.Ui;
+
+
+public class ByeCommand extends Command {
+    public boolean execute(String input, Reminder reminder, Ui ui) {
+        return false;
+    }
+}
