+++ src/main/java/Exceptions/EmptyDescriptionException.java
@@ -1,0 +1,7 @@
+package Exceptions;
+
+public class EmptyDescriptionException extends Exception {
+    public EmptyDescriptionException(String message) {
+        super(message);
+    }
+}
