+++ src/main/java/Exceptions/TooManyParametersException.java
@@ -1,0 +1,7 @@
+package Exceptions;
+
+public class TooManyParametersException extends Exception {
+    public TooManyParametersException(String message) {
+        super(message);
+    }
+}
