+++ src/main/java/Exceptions/UnsupportedCommandException.java
@@ -1,0 +1,7 @@
+package Exceptions;
+
+public class UnsupportedCommandException extends Exception {
+    public UnsupportedCommandException(String message) {
+        super(message);
+    }
+}
