+++ src/main/java/Exceptions/UnsupportedFilePathException.java
@@ -1,0 +1,7 @@
+package Exceptions;
+
+public class UnsupportedFilePathException extends Exception {
+    public UnsupportedFilePathException(String message) {
+        super(message);
+    }
+}
