+++ src/main/java/WindeTasks.txt
@@ -1,4 +1,3 @@
 D | O | cs2103 ass1 | 03/04/2026 19:00
 E | O | TAC WorldX | 05/07/2022 19:00 - 31/12/2024 18:56
-D | O | return book | 06/08/2025 18:00
-E | O | project meeting | 06/08/2025 14:00 - 06/08/2025 16:00
+D | X | return book | 06/08/2025 18:00
